Update commonGround.i
Fixed Bug discovered by Ralf Kornmann:
As some of you might know VectorBlade is using only one memory chip. Therefore all the settings and high scores will be saved on the same chip that already contains the game itself.  While the memory chip used allows to write byte by byte there is a catch. You can only change bits from one to zero. If you want to get them back to 1 you need to erase a whole 4K sector. As a state block in VectorBlade is only 80 bytes it does something smart to avoid erasing the full 4K every time it needs to save. Every time it uses another 80 block until there is no unused block left in the 4K sector. Than it needs to erase the full sector and starts over again. This happens after 51 saves operations of any persistence data. As writing or erasing the memory need some time and the chip cannot be used while the operation is in progress a part of the code needs to be copied to the RAM and executed from there. And this is were the error happens. The code first copies the function to write a byte to the flash into the RAM. Than it checks if there is still a free slot. If there is one it uses the code in the RAM to write the 80 bytes to the flash chip. But if there is no free block anymore it need to erase the full 4K first. To do this it copies the erase function in the RAM and execute it. After this it tries to write the 80 bytes but the function in the RAM is still the erase function. Instead of writing 80 bytes. It erases the 4K sector another 80 times and doesn't save anything at all. If you quit the game now it will not find any valid data the next time you start it. This cases everything is scaled to maximum and I was not able to get back in the calibration to scale it down again. If there is another save after the flash was erased it will write an valid block again and everything is fine.